% Auto-generated: do not edit by hand
\name{areaBump}

\alias{areaBump}

\title{AreaBump component}

\description{

}

\usage{
areaBump(id=NULL, axisBottom=NULL, axisTop=NULL, blendMode=NULL,
clickedPoint=NULL, colors=NULL, data=NULL, defs=NULL,
endLabel=NULL, fill=NULL, margin=NULL, spacing=NULL,
startLabel=NULL)
}

\arguments{
\item{id}{Character. }

\item{axisBottom}{Named list. }

\item{axisTop}{Named list. }

\item{blendMode}{Character. }

\item{clickedPoint}{Named list. }

\item{colors}{Named list. }

\item{data}{Unnamed list. }

\item{defs}{Unnamed list. }

\item{endLabel}{Character. }

\item{fill}{Unnamed list. }

\item{margin}{Named list. }

\item{spacing}{Numeric. }

\item{startLabel}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

