% Auto-generated: do not edit by hand
\name{responsiveCircle}

\alias{responsiveCircle}

\title{ResponsiveCircle component}

\description{

}

\usage{
responsiveCircle(id=NULL, borderColor=NULL, borderWidth=NULL,
childColor=NULL, colors=NULL, data=NULL, defs=NULL,
enableLabels=NULL, fill=NULL, labelTextColor=NULL,
labelsFilter=NULL, labelsSkipRadius=NULL, margin=NULL,
motionConfig=NULL, padding=NULL, zoomedId=NULL)
}

\arguments{
\item{id}{Character. }

\item{borderColor}{Named list. }

\item{borderWidth}{Numeric. }

\item{childColor}{Named list. }

\item{colors}{Named list. }

\item{data}{Named list. }

\item{defs}{Unnamed list. }

\item{enableLabels}{Logical. }

\item{fill}{Unnamed list. }

\item{labelTextColor}{Named list. }

\item{labelsFilter}{}

\item{labelsSkipRadius}{Numeric. }

\item{margin}{Named list. }

\item{motionConfig}{Character. }

\item{padding}{Numeric. }

\item{zoomedId}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

